refactor: Modularize code
